requirements.txt
flask==3.0.0
flask-cors==4.0.0
boto3==1.34.0
werkzeug==3.0.1
python-dotenv==1.0.0
opencv-python-headless==4.9.0.80
Pillow==10.2.0
moviepy>=1.0.3
elevenlabs>=1.0.0
requests>=2.31.0
gunicorn==21.2.0
